Move Core and Parser to TinyLang.Field.Typed
